Crate notion_core

The main implementation crate for the core of Notion.

Modules

distro
Provides types for fetching tool distributions into the local inventory.

env
Provides utilities for extracting standard Notion environment variables.

event
Events for the sessions in executables and shims and everything

fs
Provides utilities for operating on the filesystem.

hook
Provides types for working with Notion hooks.

inventory
Provides types for working with Notion's inventory, the local repository of available tool versions.

manifest
Provides the Manifest type, which represents a Node manifest file (package.json).

monitor

path
Provides functions for determining the paths of files and directories in a standard Notion layout.

platform

project
Provides the Project type, which represents a Node project tree in the filesystem.

session
Provides the Session type, which represents the user's state during an execution of a Notion tool, including their current directory, Notion hook configuration, and the state of the local inventory.

shell

shim
Provides utilities for modifying shims for 3rd-party executables

style
The view layer of Notion, with utilities for styling command-line output.

tool
Traits and types for executing command-line tools.

toolchain

version