Module notion_core::catalog

Provides types for working with Notion's local catalog, the local repository of available tool versions.

Re-exports

use std::collections::BTreeMap;

use std::collections::BTreeSet;

use std::collections::HashSet;

use std::fs::remove_dir_all;

use std::fs::File;

use std::io;

use std::io::Write;

use std::str::FromStr;

use std::string::ToString;

use lazycell::LazyCell;

use readext::ReadExt;

use reqwest;

use toml;

use config::Config;

use config::NodeConfig;

use installer::Installed;

use installer::node::Installer;

use notion_fail::Fallible;

use notion_fail::NotionError;

use notion_fail::NotionFail;

use notion_fail::ResultExt;

use path;

use path::user_catalog_file;

use semver::Version;

use semver::VersionReq;

use serial;

use serial::touch;

use style::progress_spinner;

Structs

Catalog
The catalog of tool versions available locally.

Index
The index of the public Node server.

LazyCatalog
Lazily loaded tool catalog.

NoNodeVersionFoundError
Thrown when there is no Node version matching a requested semver specifier.

NodeCatalog
The catalog of Node versions available locally.

VersionData
The set of available files on the public Node server for a given Node version.

Constants

PUBLIC_NODE_VERSION_INDEX
URL of the index of available Node versions on the public Node server.