Struct notion_core::catalog::Catalog

pub struct Catalog {
    pub node: NodeCatalog,
}

The catalog of tool versions available locally.

Fields
node: NodeCatalog
Methods

impl Catalog

fn current() -> Fallible<Catalog>

Returns the current tool catalog.

pub fn to_string(&self) -> String

Returns a pretty-printed TOML representation of the contents of the catalog.

pub fn save(&self) -> Fallible<()>

Saves the contents of the catalog to the user's catalog file.

pub fn activate_node(
&mut self,
matching: &VersionReq,
config: &Config
) -> Fallible<()>

Activates a Node version matching the specified semantic versioning requirements.

pub fn install_node(
&mut self,
matching: &VersionReq,
config: &Config
) -> Fallible<Installed>

Installs a Node version matching the specified semantic versioning requirements.

pub fn uninstall_node(&mut self, version: &Version) -> Fallible<()>

Uninstalls a specific Node version from the local catalog.

impl Catalog

pub fn to_serial(&self) -> Catalog