Struct notion_core::catalog::NodeCatalog

pub struct NodeCatalog {
    pub activated: Option<Version>,
    pub versions: BTreeSet<Version>,
}

The catalog of Node versions available locally.

Fields
activated: Option<Version>

The currently activated Node version, if any.

versions: BTreeSet<Version>
Methods

impl NodeCatalog

pub fn contains(&self, version: &Version) -> bool

Tests whether this Node catalog contains the specified Node version.

fn resolve_remote(
&self,
matching: &VersionReq,
config: &Config
) -> Fallible<Installer>

Resolves the specified semantic versioning requirements from a remote distributor.

fn resolve_public(&self, matching: &VersionReq) -> Fallible<Installer>

Resolves the specified semantic versioning requirements from the public distributor (https://nodejs.org).

pub fn resolve_local(&self, req: &VersionReq) -> Option<Version>

Resolves the specified semantic versioning requirements from the local catalog.

impl NodeCatalog

fn to_serial(&self) -> NodeCatalog