Trait notion_core::catalog::Resolve

pub trait Resolve<D: Distro> {
    fn resolve_public(&self, matching: &VersionSpec) -> Fallible<D>;

    fn resolve_remote(
        &self, 
        matching: &VersionSpec, 
        config: Option<&ToolConfig<D>>
    ) -> Fallible<D> { ... }
}

Required Methods

fn resolve_public(&self, matching: &VersionSpec) -> Fallible<D>

Resolves the specified semantic versioning requirements from the public distributor (e.g. https://nodejs.org).

Provided Methods

fn resolve_remote(
&self,
matching: &VersionSpec,
config: Option<&ToolConfig<D>>
) -> Fallible<D>

Resolves the specified semantic versioning requirements from a remote distributor.