Trait notion_core::distro::Distro

pub trait Distro: Sized {
    type VersionDetails;
    type ResolvedVersion;
    fn new(
        name: String, 
        version: Self::ResolvedVersion, 
        hooks: Option<&ToolHooks<Self>>
    ) -> Fallible<Self>;

    fn version(&self) -> &Version;

    fn fetch(
        self, 
        collection: &Collection<Self>
    ) -> Fallible<Fetched<Self::VersionDetails>>;
}

Associated Types

type VersionDetails

type ResolvedVersion

Loading content...
Required methods

fn new(
name: String,
version: Self::ResolvedVersion,
hooks: Option<&ToolHooks<Self>>
) -> Fallible<Self>

Provisions a new Distro based on the name, Version and Possible Hooks

fn version(&self) -> &Version

Produces a reference to this distro's Tool version.

fn fetch(
self,
collection: &Collection<Self>
) -> Fallible<Fetched<Self::VersionDetails>>

Fetches this version of the Tool. (It is left to the responsibility of the Collection to update its state after fetching succeeds.)

Loading content...