Enum notion_core::error::ErrorDetails

pub enum ErrorDetails {
    BinaryAlreadyInstalled {
        bin_name: String,
        package: String,
        version: String,
    },
    BinaryExecError {
        error: String,
    },
    CannotPinPackage,
    CliParseError,
    CommandNotImplemented {
        command_name: String,
    },
    CouldNotDetermineTool,
    CreateDirError {
        dir: String,
        error: String,
    },
    DepPackageReadError {
        error: String,
    },
    DeprecatedCommandError {
        command: String,
        advice: String,
    },
    DownloadToolNetworkError {
        tool: ToolSpec,
        from_url: String,
        error: String,
    },
    DownloadToolNotFound {
        tool: ToolSpec,
    },
    InvalidHookCommand {
        command: String,
    },
    NoBinPlatform {
        binary: String,
    },
    NodeVersionNotFound {
        matching: String,
    },
    NoGlobalInstalls,
    NoHomeEnvironmentVar,
    NoLocalDataDir,
    NoPackageExecutables,
    NoPackageFound {
        name: String,
        matching: VersionSpec,
    },
    NoPinnedNodeVersion,
    NoSuchTool {
        tool: String,
    },
    NotInPackage,
    NoToolChain {
        shim_name: String,
    },
    NoVersionsFound,
    NpxNotAvailable {
        version: String,
    },
    PackageInstallFailed {
        cmd: String,
        status: ExitStatus,
    },
    PackageInstallIoError {
        error: String,
    },
    PackageReadError {
        error: String,
    },
    PackageUnpackError,
    PathError,
    RegistryFetchError {
        error: String,
    },
    SymlinkError {
        error: String,
    },
    ToolNotImplemented,
    UnrecognizedShell {
        name: String,
    },
    UnspecifiedPostscript,
    UnspecifiedShell,
    VersionParseError {
        error: String,
    },
    YarnVersionNotFound {
        matching: String,
    },
}

Variants
BinaryAlreadyInstalled

Thrown when package tries to install a binary that is already installed.

Fields of BinaryAlreadyInstalled

bin_name: Stringpackage: Stringversion: String
BinaryExecError
Fields of BinaryExecError

error: String
CannotPinPackage

Thrown when a user tries to notion pin something other than node/yarn/npm.

CliParseErrorCommandNotImplemented
Fields of CommandNotImplemented

command_name: String
CouldNotDetermineToolCreateDirError
Fields of CreateDirError

dir: Stringerror: String
DepPackageReadError
Fields of DepPackageReadError

error: String
DeprecatedCommandError
Fields of DeprecatedCommandError

command: Stringadvice: String
DownloadToolNetworkError
Fields of DownloadToolNetworkError

tool: ToolSpecfrom_url: Stringerror: String
DownloadToolNotFound
Fields of DownloadToolNotFound

tool: ToolSpec
InvalidHookCommand
Fields of InvalidHookCommand

command: String
NoBinPlatform

Thrown when BinConfig (read from file) does not contain Platform info.

Fields of NoBinPlatform

binary: String
NodeVersionNotFound

Thrown when there is no Node version matching a requested semver specifier.

Fields of NodeVersionNotFound

matching: String
NoGlobalInstallsNoHomeEnvironmentVarNoLocalDataDirNoPackageExecutables

Thrown when a user tries to install or fetch a package with no executables.

NoPackageFound

Thrown when there is no package version matching a requested semver specifier.

Fields of NoPackageFound

name: Stringmatching: VersionSpec
NoPinnedNodeVersion

Thrown when a user tries to pin a Yarn version before pinning a Node version.

NoSuchTool
Fields of NoSuchTool

tool: String
NotInPackage

Thrown when the user tries to pin Node or Yarn versions outside of a package.

NoToolChain
Fields of NoToolChain

shim_name: String
NoVersionsFoundNpxNotAvailable
Fields of NpxNotAvailable

version: String
PackageInstallFailed

Thrown when package install command is not successful.

Fields of PackageInstallFailed

cmd: Stringstatus: ExitStatus
PackageInstallIoError

Thrown when package install command fails to execute.

Fields of PackageInstallIoError

error: String
PackageReadError
Fields of PackageReadError

error: String
PackageUnpackError

Thrown when a package has been unpacked but is not formed correctly.

PathErrorRegistryFetchError

Thrown when the public registry for Node or Yarn could not be downloaded.

Fields of RegistryFetchError

error: String
SymlinkError
Fields of SymlinkError

error: String
ToolNotImplementedUnrecognizedShell

Thrown when the shell name specified in the Notion environment is not supported.

Fields of UnrecognizedShell

name: String
UnspecifiedPostscript

Thrown when the postscript file was not specified in the Notion environment.

UnspecifiedShell

Thrown when the shell name was not specified in the Notion environment.

VersionParseError
Fields of VersionParseError

error: String
YarnVersionNotFound

Thrown when there is no Yarn version matching a requested semver specifier.

Fields of YarnVersionNotFound

matching: String