Struct notion_core::installer::node::Installer

pub struct Installer {
    archive: Box<Archive>,
    version: Version,
}

A provisioned Node installer.

Fields
archive: Box<Archive> version: Version
Methods

impl Installer

pub fn public(version: Version) -> Fallible<Self>

Provision an Installer from the public Node distributor (https://nodejs.org).

pub fn remote(version: Version, url: &str) -> Fallible<Self>

Provision an Installer from a remote distributor.

pub fn cached(version: Version, file: File) -> Fallible<Self>

Provision an Installer from the filesystem.

pub fn version(&self) -> &Version

Produces a reference to this installer's Node version.

pub fn install(self, catalog: &NodeCatalog) -> Fallible<Installed>

Installs this version of Node. (It is left to the responsibility of the NodeCatalog to update its state after installation succeeds.)