Trait notion_core::inventory::FetchResolve

pub trait FetchResolve<D: Distro> {
    fn fetch(
        &mut self, 
        matching: &VersionSpec, 
        config: &Config
    ) -> Fallible<Fetched<DistroVersion>>;

    fn resolve_public(&self, matching: &VersionSpec) -> Fallible<D>;

    fn resolve_remote(
        &self, 
        matching: &VersionSpec, 
        config: Option<&ToolConfig<D>>
    ) -> Fallible<D> { ... }
}

Required methods

fn fetch(
&mut self,
matching: &VersionSpec,
config: &Config
) -> Fallible<Fetched<DistroVersion>>

Fetch a Distro version matching the specified semantic versioning requirements.

fn resolve_public(&self, matching: &VersionSpec) -> Fallible<D>

Resolves the specified semantic versioning requirements from the public distributor (e.g. https://nodejs.org).

Loading content...
Provided methods

fn resolve_remote(
&self,
matching: &VersionSpec,
config: Option<&ToolConfig<D>>
) -> Fallible<D>

Resolves the specified semantic versioning requirements from a remote distributor.

Loading content...