Module notion_core::path

Provides functions for determining the paths of files and directories in a standard Notion layout.

Modules

unix	Unix
Provides functions for determining the paths of files and directories in a standard Notion layout in Unix-based operating systems.

windows	Windows
Provides functions for determining the paths of files and directories in a standard Notion layout in Windows operating systems.

Constants

ARCH
The system architecture component of a Node distribution tarball's name.

OS
The OS component of a Node distribution tarball's name.

Functions

archive_extension

cache_dir

create_file_symlink

default_notion_home

image_dir

inventory_dir

launchbin_file

launchscript_file

node_archive_npm_package_json_path

node_archive_root_dir_name

node_cache_dir

node_distro_file_name

node_image_3p_bin_dir

node_image_bin_dir

node_image_dir

node_image_root_dir

node_index_expiry_file

node_index_file

node_inventory_dir

node_npm_version_file

notion_file

notion_home

package_inventory_dir

shim_dir

shim_file

tmp_dir

tools_dir

user_hooks_file

user_platform_file

user_toolchain_dir

yarn_archive_root_dir_name

yarn_distro_file_name

yarn_image_bin_dir

yarn_image_dir

yarn_image_root_dir

yarn_inventory_dir