Struct notion_core::project::Project

pub struct Project {
    manifest: Manifest,
    project_root: PathBuf,
    dependent_bins: LazyDependentBins,
}

A Node project tree in the filesystem.

Fields
manifest: Manifestproject_root: PathBufdependent_bins: LazyDependentBins
Methods

impl Project

fn for_current_dir() -> Fallible<Option<Rc<Project>>>

Returns the Node project containing the current working directory, if any.

fn for_dir(base_dir: &Path) -> Fallible<Option<Rc<Project>>>

Returns the Node project for the input directory, if any.

pub fn platform(&self) -> Option<Rc<PlatformSpec>>

Returns the pinned platform image, if any.

pub fn is_pinned(&self) -> bool

Returns true if the project manifest contains a toolchain.

pub fn manifest(&self) -> &Manifest

Returns the project manifest (package.json) for this project.

pub fn package_file(&self) -> PathBuf

Returns the path to the package.json file for this project.

pub fn local_bin_dir(&self) -> PathBuf

Returns the path to the local binary directory for this project.

pub fn has_direct_bin(&self, bin_name: &OsStr) -> Fallible<bool>

Returns true if the input binary name is a direct dependency of the input project

fn dependent_binaries(&self) -> Fallible<HashMap<String, String>>

Returns a mapping of the names to paths for all the binaries installed by direct dependencies of the current project.

fn get_dependency_path(&self, name: &String) -> PathBuf

Convert dependency names to the path to each project.

pub fn pin_node(&self, node_version: &NodeVersion) -> Fallible<()>

Writes the specified version of Node to the toolchain.node key in package.json.

pub fn pin_yarn(&self, yarn_version: &Version) -> Fallible<()>

Writes the specified version of Yarn to the toolchain.yarn key in package.json.

pub fn pin_npm(&self, npm_version: &Version) -> Fallible<()>

Writes the specified version of Npm to the toolchain.npm key in package.json.