Struct notion_core::session::Session

pub struct Session {
    hooks: LazyHookConfig,
    inventory: LazyInventory,
    toolchain: LazyToolchain,
    project: LazyProject,
    event_log: EventLog,
}

Represents the user's state during an execution of a Notion tool. The session encapsulates a number of aspects of the environment in which the tool was invoked, including: - the current directory - the Node project tree that contains the current directory (if any) - the Notion hook configuration - the inventory of locally-fetched Notion tools

Fields
hooks: LazyHookConfiginventory: LazyInventorytoolchain: LazyToolchainproject: LazyProjectevent_log: EventLog
Methods

impl Session

pub fn new() -> Session

Constructs a new Session.

pub fn project(&self) -> Fallible<Option<Rc<Project>>>

Produces a reference to the current Node project, if any.

pub fn current_platform(&self) -> Fallible<Option<Rc<PlatformSpec>>>

pub fn user_platform(&self) -> Fallible<Option<Rc<PlatformSpec>>>

pub fn project_platform(&self) -> Fallible<Option<Rc<PlatformSpec>>>

Returns the current project's pinned platform image, if any.

pub fn inventory(&self) -> Fallible<&Inventory>

Produces a reference to the current inventory.

pub fn inventory_mut(&mut self) -> Fallible<&mut Inventory>

Produces a mutable reference to the current inventory.

pub fn hooks(&self) -> Fallible<&HookConfig>

Produces a reference to the hook configuration

pub(crate) fn ensure_node(&mut self, version: &Version) -> Fallible<()>

Ensures that a specific Node version has been fetched and unpacked

pub(crate) fn ensure_yarn(&mut self, version: &Version) -> Fallible<()>

Ensures that a specific Yarn version has been fetched and unpacked

pub fn install_node(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Fetch and unpack a version of Node matching the input requirements.

pub fn install_yarn(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Fetch and unpack a version of Yarn matching the input requirements.

pub fn install_npm(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Fetch, unpack, and install a version of Npm matching the input requirements.

pub fn install_package(
&mut self,
name: String,
version: &VersionSpec
) -> Fallible<Version>

Fetch, unpack, and install a package matching the input requirements.

pub fn fetch_node(
&mut self,
version_spec: &VersionSpec
) -> Fallible<Fetched<NodeVersion>>

Fetches a Node version matching the specified semantic versioning requirements.

pub fn fetch_yarn(
&mut self,
version_spec: &VersionSpec
) -> Fallible<Fetched<Version>>

Fetches a Yarn version matching the specified semantic versioning requirements.

pub fn fetch_npm(
&mut self,
version_spec: &VersionSpec
) -> Fallible<Fetched<PackageVersion>>

Fetches a Npm version matching the specified semantic versioning requirements.

pub fn fetch_package(
&mut self,
name: String,
version_spec: &VersionSpec
) -> Fallible<Fetched<PackageVersion>>

Fetches a Package version matching the specified semantic versioning requirements.

pub fn pin_node(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Updates toolchain in package.json with the Node version matching the specified semantic versioning requirements.

pub fn pin_yarn(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Updates toolchain in package.json with the Yarn version matching the specified semantic versioning requirements.

pub fn pin_npm(&mut self, version_spec: &VersionSpec) -> Fallible<()>

Updates toolchain in package.json with the Npm version matching the specified semantic versioning requirements.

pub fn get_user_tool(&mut self, tool_name: &OsStr) -> Fallible<Option<UserTool>>

Gets the installed UserTool with the input name, if any.

pub fn add_event_start(&mut self, activity_kind: ActivityKind)

pub fn add_event_end(
&mut self,
activity_kind: ActivityKind,
exit_code: ExitCode
)

pub fn add_event_tool_end(
&mut self,
activity_kind: ActivityKind,
exit_code: i32
)

pub fn add_event_error(
&mut self,
activity_kind: ActivityKind,
error: &NotionError
)

fn publish_to_event_log(self)

pub fn exit(self, code: ExitCode) -> !

pub fn exit_tool(self, code: i32) -> !