Module notion_core::catalog

Provides types for working with Notion's local catalog, the local repository of available tool versions.

Modules

serial

Structs

Catalog
The catalog of tool versions available locally.

Collection

Index
The index of the public Node server.

LazyCatalog
Lazily loaded tool catalog.

NoNodeVersionFoundError
Thrown when there is no Node version matching a requested semver specifier.

NoYarnVersionFoundError
Thrown when there is no Yarn version matching a requested semver specifier.

RegistryFetchError
Thrown when the public registry for Node or Yarn could not be downloaded.

VersionData
The set of available files on the public Node server for a given Node version.

Traits

Resolve

Functions

max_age
Get the cache max-age of an HTTP reponse.

public_node_version_index
Returns the URL of the index of available Node versions on the public Node server.

public_yarn_latest_version
URL of the latest Yarn version on the public yarnpkg.com

public_yarn_version_index
Return the URL of the index of available Yarn versions on the public git repository.

read_cached_opt
Reads a public index from the Node cache, if it exists and hasn't expired.

resolve_node_versions

Type Definitions

NodeCollection

YarnCollection