Trait notion_core::distro::Distro

pub trait Distro: Sized {
    type VersionDetails;
    fn new(version: Version, hooks: Option<&ToolHooks<Self>>) -> Fallible<Self>;

    fn version(&self) -> &Version;

    fn fetch(
        self, 
        collection: &Collection<Self>
    ) -> Fallible<Fetched<DistroVersion>>;
}

Associated Types

type VersionDetails

Loading content...
Required methods

fn new(version: Version, hooks: Option<&ToolHooks<Self>>) -> Fallible<Self>

Provisions a new Distro based on the Version and Possible Hooks

fn version(&self) -> &Version

Produces a reference to this distro's Tool version.

fn fetch(
self,
collection: &Collection<Self>
) -> Fallible<Fetched<DistroVersion>>

Fetches this version of the Tool. (It is left to the responsibility of the Collection to update its state after fetching succeeds.)

Loading content...