Enum notion_core::error::ErrorDetails

pub enum ErrorDetails {
    BinaryAlreadyInstalled {
        bin_name: String,
        existing_package: String,
        new_package: String,
    },
    BinaryExecError,
    BinaryNotFound {
        name: String,
    },
    BuildPathError,
    CannotPinPackage {
        package: String,
    },
    CompletionsOutDirError,
    ContainingDirError {
        path: String,
    },
    CouldNotDetermineTool,
    CreateDirError {
        dir: String,
    },
    CreatePostscriptError {
        in_dir: String,
    },
    CreateTempDirError {
        in_dir: String,
    },
    CreateTempFileError {
        in_dir: String,
    },
    CurrentDirError,
    DeleteDirectoryError {
        directory: String,
    },
    DeleteFileError {
        file: String,
    },
    DepPackageReadError,
    DeprecatedCommandError {
        command: String,
        advice: String,
    },
    DownloadToolNetworkError {
        tool: ToolSpec,
        from_url: String,
    },
    ExecutablePathError {
        command: String,
    },
    ExecuteHookError {
        command: String,
    },
    HookMultipleFieldsSpecified,
    HookNoFieldsSpecified,
    InvalidHookCommand {
        command: String,
    },
    InvalidHookOutput {
        command: String,
    },
    NoBinPlatform {
        binary: String,
    },
    NodeVersionNotFound {
        matching: String,
    },
    NoGlobalInstalls,
    NoHomeEnvironmentVar,
    NoLocalDataDir,
    NoPackageExecutables,
    NoPinnedNodeVersion,
    NoPlatform,
    NoProjectYarn,
    NotInPackage,
    NoUserYarn,
    NoVersionsFound,
    NpxNotAvailable {
        version: String,
    },
    PackageInstallFailed,
    PackageMetadataFetchError {
        from_url: String,
    },
    PackageParseError {
        file: String,
    },
    PackageReadError {
        file: String,
    },
    PackageUnpackError,
    PackageVersionNotFound {
        name: String,
        matching: String,
    },
    PackageWriteError {
        file: String,
    },
    ParseBinConfigError,
    ParseHooksError,
    ParseNodeIndexCacheError,
    ParseNodeIndexError {
        from_url: String,
    },
    ParseNodeIndexExpiryError,
    ParseNpmManifestError,
    ParsePackageConfigError,
    ParsePackageMetadataError {
        from_url: String,
    },
    ParsePlatformError,
    PublishHookBothUrlAndBin,
    PublishHookNeitherUrlNorBin,
    ReadBinConfigDirError {
        dir: String,
    },
    ReadBinConfigError {
        file: String,
    },
    ReadDefaultNpmError {
        file: String,
    },
    ReadHooksError {
        file: String,
    },
    ReadInventoryDirError {
        dir: String,
    },
    ReadNodeIndexCacheError {
        file: String,
    },
    ReadNodeIndexExpiryError {
        file: String,
    },
    ReadNpmManifestError,
    ReadPackageConfigError {
        file: String,
    },
    ReadPlatformError {
        file: String,
    },
    RegistryFetchError {
        tool: String,
        from_url: String,
    },
    SetupToolImageError {
        tool: String,
        version: String,
        dir: String,
    },
    ShimCreateError {
        name: String,
    },
    ShimRemoveBuiltInError {
        name: String,
    },
    ShimRemoveError {
        name: String,
    },
    StringifyBinConfigError,
    StringifyPackageConfigError,
    StringifyPlatformError,
    StringifyToolchainError,
    UnpackArchiveError {
        tool: String,
        version: String,
    },
    UnrecognizedShell {
        name: String,
    },
    UnspecifiedPostscript,
    UnspecifiedShell,
    VersionParseError {
        version: String,
    },
    WriteBinConfigError {
        file: String,
    },
    WriteDefaultNpmError {
        file: String,
    },
    WriteNodeIndexCacheError {
        file: String,
    },
    WriteNodeIndexExpiryError {
        file: String,
    },
    WritePackageConfigError {
        file: String,
    },
    WritePackageShasumError {
        package: String,
        version: String,
        file: String,
    },
    WritePlatformError {
        file: String,
    },
    YarnLatestFetchError {
        from_url: String,
    },
    YarnVersionNotFound {
        matching: String,
    },
}

Variants
BinaryAlreadyInstalled

Thrown when package tries to install a binary that is already installed.

Fields of BinaryAlreadyInstalled

bin_name: Stringexisting_package: Stringnew_package: String
BinaryExecErrorBinaryNotFound

Thrown when a binary could not be found in the local inventory

Fields of BinaryNotFound

name: String
BuildPathError

Thrown when building the virtual environment path fails

CannotPinPackage

Thrown when a user tries to notion pin something other than node/yarn/npm.

Fields of CannotPinPackage

package: String
CompletionsOutDirError

Thrown when the Completions out-dir is not a directory

ContainingDirError

Thrown when the containing directory could not be determined

Fields of ContainingDirError

path: String
CouldNotDetermineToolCreateDirError
Fields of CreateDirError

dir: String
CreatePostscriptError

Thrown when unable to create the postscript file

Fields of CreatePostscriptError

in_dir: String
CreateTempDirError

Thrown when creating a temporary directory fails

Fields of CreateTempDirError

in_dir: String
CreateTempFileError

Thrown when creating a temporary file fails

Fields of CreateTempFileError

in_dir: String
CurrentDirErrorDeleteDirectoryError

Thrown when deleting a directory fails

Fields of DeleteDirectoryError

directory: String
DeleteFileError

Thrown when deleting a file fails

Fields of DeleteFileError

file: String
DepPackageReadError

Thrown when reading dependency package info fails

DeprecatedCommandError
Fields of DeprecatedCommandError

command: Stringadvice: String
DownloadToolNetworkError
Fields of DownloadToolNetworkError

tool: ToolSpecfrom_url: String
ExecutablePathError

Thrown when building the path to an executable fails

Fields of ExecutablePathError

command: String
ExecuteHookError

Thrown when executing a hook command fails

Fields of ExecuteHookError

command: String
HookMultipleFieldsSpecified

Thrown when a hook contains multiple fields (prefix, template, or bin)

HookNoFieldsSpecified

Thrown when a hook doesn't contain any of the known fields (prefix, template, or bin)

InvalidHookCommand
Fields of InvalidHookCommand

command: String
InvalidHookOutput

Thrown when output from a hook command could not be read

Fields of InvalidHookOutput

command: String
NoBinPlatform

Thrown when BinConfig (read from file) does not contain Platform info.

Fields of NoBinPlatform

binary: String
NodeVersionNotFound

Thrown when there is no Node version matching a requested semver specifier.

Fields of NodeVersionNotFound

matching: String
NoGlobalInstallsNoHomeEnvironmentVarNoLocalDataDirNoPackageExecutables

Thrown when a user tries to install or fetch a package with no executables.

NoPinnedNodeVersion

Thrown when a user tries to pin a Yarn version before pinning a Node version.

NoPlatform

Thrown when the platform (Node version) could not be determined

NoProjectYarn

Thrown when Yarn is not set in a project

NotInPackage

Thrown when the user tries to pin Node or Yarn versions outside of a package.

NoUserYarn

Thrown when default Yarn is not set

NoVersionsFoundNpxNotAvailable
Fields of NpxNotAvailable

version: String
PackageInstallFailed

Thrown when package install command is not successful.

PackageMetadataFetchError

Thrown when there is an error fetching package metadata

Fields of PackageMetadataFetchError

from_url: String
PackageParseError

Thrown when parsing a package manifest fails

Fields of PackageParseError

file: String
PackageReadError

Thrown when reading a package manifest fails

Fields of PackageReadError

file: String
PackageUnpackError

Thrown when a package has been unpacked but is not formed correctly.

PackageVersionNotFound

Thrown when there is no package version matching a requested semver specifier.

Fields of PackageVersionNotFound

name: Stringmatching: String
PackageWriteError

Thrown when writing a package manifest fails

Fields of PackageWriteError

file: String
ParseBinConfigError

Thrown when unable to parse a bin config file

ParseHooksError

Thrown when unable to parse the hooks.toml file

ParseNodeIndexCacheError

Thrown when unable to parse the node index cache

ParseNodeIndexError

Thrown when unable to parse the node index

Fields of ParseNodeIndexError

from_url: String
ParseNodeIndexExpiryError

Thrown when unable to parse the node index cache expiration

ParseNpmManifestError

Thrown when unable to parse the npm manifest file from a node install

ParsePackageConfigError

Thrown when unable to parse a package configuration

ParsePackageMetadataError

Thrown when unable to parse the metadata for a package

Fields of ParsePackageMetadataError

from_url: String
ParsePlatformError

Thrown when unable to parse the platform.json file

PublishHookBothUrlAndBin

Thrown when a publish hook contains both the url and bin fields

PublishHookNeitherUrlNorBin

Thrown when a publish hook contains neither url nor bin fields

ReadBinConfigDirError

Thrown when there was an error reading the user bin directory

Fields of ReadBinConfigDirError

dir: String
ReadBinConfigError

Thrown when there was an error reading the config for a binary

Fields of ReadBinConfigError

file: String
ReadDefaultNpmError

Thrown when unable to read the default npm version file

Fields of ReadDefaultNpmError

file: String
ReadHooksError

Thrown when there was an error opening the hooks.toml file

Fields of ReadHooksError

file: String
ReadInventoryDirError

Thrown when there was an error reading the inventory contents

Fields of ReadInventoryDirError

dir: String
ReadNodeIndexCacheError

Thrown when there was an error reading the Node Index Cache

Fields of ReadNodeIndexCacheError

file: String
ReadNodeIndexExpiryError

Thrown when there was an error reading the Node Index Cache Expiration

Fields of ReadNodeIndexExpiryError

file: String
ReadNpmManifestError

Thrown when there was an error reading the npm manifest file

ReadPackageConfigError

Thrown when there was an error reading a package configuration file

Fields of ReadPackageConfigError

file: String
ReadPlatformError

Thrown when there was an error opening the user platform file

Fields of ReadPlatformError

file: String
RegistryFetchError

Thrown when the public registry for Node or Yarn could not be downloaded.

Fields of RegistryFetchError

tool: Stringfrom_url: String
SetupToolImageError

Thrown when there was an error copying an unpacked tool to the image directory

Fields of SetupToolImageError

tool: Stringversion: Stringdir: String
ShimCreateError

Thrown when Notion is unable to create a shim

Fields of ShimCreateError

name: String
ShimRemoveBuiltInError

Thrown when trying to remove a built-in shim (node, yarn, etc.)

Fields of ShimRemoveBuiltInError

name: String
ShimRemoveError

Thrown when Notion is unable to remove a shim

Fields of ShimRemoveError

name: String
StringifyBinConfigError

Thrown when serializnig a bin config to JSON fails

StringifyPackageConfigError

Thrown when serializnig a package config to JSON fails

StringifyPlatformError

Thrown when serializing the platform to JSON fails

StringifyToolchainError

Thrown when serializing the toolchain to JSON fails

UnpackArchiveError

Thrown when unpacking an archive (tarball or zip) fails

Fields of UnpackArchiveError

tool: Stringversion: String
UnrecognizedShell

Thrown when the shell name specified in the Notion environment is not supported.

Fields of UnrecognizedShell

name: String
UnspecifiedPostscript

Thrown when the postscript file was not specified in the Notion environment.

UnspecifiedShell

Thrown when the shell name was not specified in the Notion environment.

VersionParseError
Fields of VersionParseError

version: String
WriteBinConfigError

Thrown when there was an error writing a bin config file

Fields of WriteBinConfigError

file: String
WriteDefaultNpmError

Thrown when there was an error writing the default npm to file

Fields of WriteDefaultNpmError

file: String
WriteNodeIndexCacheError

Thrown when there was an error writing the node index cache

Fields of WriteNodeIndexCacheError

file: String
WriteNodeIndexExpiryError

Thrown when there was an error writing the node index expiration

Fields of WriteNodeIndexExpiryError

file: String
WritePackageConfigError

Thrown when there was an error writing a package config

Fields of WritePackageConfigError

file: String
WritePackageShasumError

Thrown when there was an error writing the shasum for a package

Fields of WritePackageShasumError

package: Stringversion: Stringfile: String
WritePlatformError

Thrown when writing the platform.json file fails

Fields of WritePlatformError

file: String
YarnLatestFetchError

Thrown when there is an error fetching the latest version of Yarn

Fields of YarnLatestFetchError

from_url: String
YarnVersionNotFound

Thrown when there is no Yarn version matching a requested semver specifier.

Fields of YarnVersionNotFound

matching: String