Struct notion_core::session::Session

pub struct Session {
    config: LazyConfig,
    catalog: LazyCatalog,
    project: Option<Project>,
}

Represents the user's state during an execution of a Notion tool. The session encapsulates a number of aspects of the environment in which the tool was invoked, including: - the current directory - the Node project tree that contains the current directory (if any) - the Notion configuration settings - the catalog of locally-installed Notion tools

Fields
config: LazyConfig catalog: LazyCatalog project: Option<Project>
Methods

impl Session

pub fn new() -> Fallible<Session>

Constructs a new Session.

pub fn project(&self) -> Option<&Project>

Produces a reference to the current Node project, if any.

pub fn catalog(&self) -> Fallible<&Catalog>

Produces a reference to the current tool catalog.

pub fn catalog_mut(&mut self) -> Fallible<&mut Catalog>

Produces a mutable reference to the current tool catalog.

pub fn config(&self) -> Fallible<&Config>

Produces a reference to the configuration.

pub fn current_node(&mut self) -> Fallible<Option<Version>>

Produces the version of Node for the current session. If there is an active project with Notion settings, this will ensure a compatible version of Node is installed before returning. If there is no active project with Notion settings, this produces the global version, which may be None.

pub fn install_node(&mut self, matching: &VersionReq) -> Fallible<Installed>

Installs a version of Node matching the specified semantic verisoning requirements.

pub fn activate_node(&mut self, matching: &VersionReq) -> Fallible<()>

Activates a version of Node matching the specified semantic versioning requirements.