Struct notion_core::session::Session

pub struct Session {
    config: LazyConfig,
    catalog: LazyCatalog,
    project: Option<Rc<Project>>,
    event_log: EventLog,
}

Represents the user's state during an execution of a Notion tool. The session encapsulates a number of aspects of the environment in which the tool was invoked, including: - the current directory - the Node project tree that contains the current directory (if any) - the Notion configuration settings - the catalog of locally-installed Notion tools

Fields
config: LazyConfig catalog: LazyCatalog project: Option<Rc<Project>> event_log: EventLog
Methods

impl Session

pub fn new() -> Fallible<Session>

Constructs a new Session.

pub fn project(&self) -> Option<Rc<Project>>

Produces a reference to the current Node project, if any.

pub fn current_platform(&mut self) -> Fallible<Option<Rc<Image>>>

pub fn user_platform(&mut self) -> Fallible<Option<Rc<Image>>>

pub fn project_platform(&self) -> Option<Rc<Image>>

Returns the current project's pinned platform image, if any.

pub fn catalog(&self) -> Fallible<&Catalog>

Produces a reference to the current tool catalog.

pub fn catalog_mut(&mut self) -> Fallible<&mut Catalog>

Produces a mutable reference to the current tool catalog.

pub fn config(&self) -> Fallible<&Config>

Produces a reference to the configuration.

pub(crate) fn prepare_image(&mut self, image: &Image) -> Fallible<()>

Ensures that a platform image has been fully fetched and set up.

pub fn user_node(&self) -> Fallible<Option<Version>>

pub fn fetch_node(&mut self, matching: &VersionSpec) -> Fallible<Fetched>

Fetches a version of Node matching the specified semantic verisoning requirements.

pub fn set_user_node(&mut self, matching: &VersionSpec) -> Fallible<()>

Sets the user toolchain's Node version to one matching the specified semantic versioning requirements.

pub fn get_matching_node(&self, matching: &VersionSpec) -> Fallible<Version>

Returns the version of Node matching the specified semantic versioning requirements.

pub fn pin_node_version(&self, matching: &VersionSpec) -> Fallible<()>

Updates toolchain in package.json with the Node version matching the specified semantic versioning requirements.

pub fn user_yarn(&mut self) -> Fallible<Option<Version>>

pub fn fetch_yarn(&mut self, matching: &VersionSpec) -> Fallible<Fetched>

Fetches a version of Node matching the specified semantic verisoning requirements.

pub fn set_user_yarn(&mut self, matching: &VersionSpec) -> Fallible<()>

Sets the Yarn version in the user toolchain to one matching the specified semantic versioning requirements.

pub fn get_matching_yarn(&self, matching: &VersionSpec) -> Fallible<Version>

Returns the version of Yarn matching the specified semantic versioning requirements

pub fn pin_yarn_version(&self, matching: &VersionSpec) -> Fallible<()>

Updates toolchain in package.json with the Yarn version matching the specified semantic versioning requirements.

pub fn add_event_start(&mut self, activity_kind: ActivityKind)

pub fn add_event_end(
&mut self,
activity_kind: ActivityKind,
exit_code: ExitCode
)

pub fn add_event_tool_end(
&mut self,
activity_kind: ActivityKind,
exit_code: i32
)

pub fn add_event_error(
&mut self,
activity_kind: ActivityKind,
error: &NotionError
)

fn publish_to_event_log(self)

pub fn exit(self, code: ExitCode) -> !

pub fn exit_tool(self, code: i32) -> !